+++ RELEASE-NOTES.md
@@ -1,3 +1,4 @@
+
 0.13.0 (Not yet released)
 ------
 
@@ -55,6 +56,8 @@
     - `addInstances`
     - `replaceCronTemplate`
 - Task ID strings are no longer prefixed by a timestamp.
+- The scheduler previously supported specification of command line arguments by fully-qualified
+  class names.  This support has been removed.
 
 0.12.0
 ------